Try to guess a default value for Git e-mail :spades:
The script tries to get a default e-mail from the `git log` of the repo.
If you made any change to the repository at any time of its history, and
your linux user has propperly set your name (or even part of it), there
is a big chance of getting a correct guess for your Git e-mail.