Fix prompt colors so spaces are properly calculated :space_invader:
This change moves the PS1 definition away from set_prompt() and instead leaves
its creation to a function, which is invoked every time the prompt is printed
(`PROMPT_COMMAND`). This allowed the usage of `\[...\]`, yielding a reliable
char count for Bash.

A bug where home paths compressed (using "…/") hided the home icon was solved,
and the path length calculation now accounts for the prefix length too.