Set Git's default push locally :sweat:
Old Git versions don't support the `simple` method for `push.default`,
so as it was before (with this key defined in `.gitconfig`) it was
causing errors for old versions. This change moves the setting to the
`.gitconfig.local` file, adding the key on its bundle installation.

It's definitely not a fix but a workaround, as it would still require
manually removing the `.gitconfig.local` definition of `push.default`
if it fails. Blame it on Git and its untolerant handling of the config
files.