Another attempt at making GitHub actions work :upside_down_face:
The shellcheck task will remain as a warning until I have time
to go over all the errors and fix them.

In the future, I'd like it to run only against changed files,
and move the Bats "links" test to another workflow so it doesn't
break the tests for unrelated changes.

PS: If you are reading this, thanks for investing your time in
my dotfiles, hope you find them interesting :)